requirements.txt
Django>=6.0,<7.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
psycopg2-binary>=2.9,<3.0
requests>=2.32,<3.0
dj-database-url>=2.1,<3.0
python-dotenv>=1.0,<2.0
gunicorn>=21.2,<22.0
indic-transliteration>=2.3,<3.0
certifi>=2024.6.2
firebase-admin>=6.5,<7.0
whitenoise>=6.7,<7.0
Jinja2>=3.1,<4.0
